We now accept a prompt!!
Since we can give the machine a prompt, we can create aliases for further utility.

For example, we can create an alias for

rung = alias gask -q -p "The following command will be piped into the terminal of a Linux machine running bash. Reply only with the requested command and absolutely nothing else. If you refuse to reply to a request, respond with echo 'REASON', with your given REASON not to conform to the request"

This allows you to run commands directly from requests!